requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
discord.py==2.4.0
discord.py[voice]==2.4.0
fastapi==0.110.2
ffmpeg==1.4
Flask==3.0.3
Flask-SocketIO==5.3.6
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hiredis==2.3.2
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
openai==1.23.2
packaging==24.0
priority==2.0.0
PyAudio==0.2.14
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
pyee==11.1.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-ffmpeg==2.0.12
python-multipart==0.0.9
python-socketio==5.11.2
requests==2.32.3
scipy==1.13.1
setuptools==69.5.1
simple-websocket==1.0.0
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.2
typing_extensions==4.10.0
websockets==12.0
Werkzeug==3.0.2
wsproto==1.2.0
